requirements.txt
arrow==0.13.1
binaryornot==0.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
Django==2.2
django-json-field==0.5.7
django-shell-plus==1.1.7
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.5
future==0.17.1
gunicorn==19.9.0
idna==2.8
Jinja2==2.10.1
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.4.2
psycopg2==2.7.4
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
whichcraft==0.5.2
whitenoise==4.1.2
